travel/requirements.txt
# FastAPI 和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 现有项目依赖
requests==2.31.0
langchain-core==0.1.0
langchain-openai==0.0.5
langgraph==0.0.26
python-dotenv==1.0.0
json-repair==0.7.0

# 高德地图API
amap-python==1.0.0

# 其他工具
markdown-it-py==3.0.0
